Get rid of SitemapIndex for SitemapRepository
