+++ src/GraphQL/Fields/SitemapField.php
@@ -2,13 +2,13 @@
 
 namespace Aerni\AdvancedSeo\GraphQL\Fields;
 
-use Aerni\AdvancedSeo\GraphQL\Types\SeoSitemapType;
-use Aerni\AdvancedSeo\Sitemaps\SitemapIndex;
-use GraphQL\Type\Definition\ResolveInfo;
-use GraphQL\Type\Definition\Type;
-use Illuminate\Support\Collection;
-use Rebing\GraphQL\Support\Field;
 use Statamic\Facades\GraphQL;
+use GraphQL\Type\Definition\Type;
+use Rebing\GraphQL\Support\Field;
+use Illuminate\Support\Collection;
+use Aerni\AdvancedSeo\Facades\Sitemap;
+use GraphQL\Type\Definition\ResolveInfo;
+use Aerni\AdvancedSeo\GraphQL\Types\SeoSitemapType;
 
 class SitemapField extends Field
 {
@@ -42,7 +42,7 @@ class SitemapField extends Field
 
     public function resolve($root, $args, $context, ResolveInfo $info): ?Collection
     {
-        $sitemaps = (new SitemapIndex)->{"{$info->fieldName}Sitemaps"}();
+        $sitemaps = Sitemap::{"{$info->fieldName}Sitemaps"}();
 
         if ($baseUrl = $args['baseUrl'] ?? null) {
             $sitemaps = $sitemaps->each(fn ($sitemap) => $sitemap->baseUrl($baseUrl));
